Update Foundation to v. 6.2.3
The previous boilerplate, using foundation 6.1.1 failed when using production build step. Some of the 6.1.1 javascript modules defined themselves as amd modules - but because the core foundation functionality did not, the packaged file failed.

Also, the current Foundation 6 does not support loading modules with RequireJS. So this update removes the possibility to load individual files, and instead includes the full foundation library always.

This is meant as a quick fix so that the boilerplate works, also in production builds