Add grunt configuration for building a Umbraco Styling Plugin
The task combines and builds all .scss files that ends with .umbraco.scss into one file, and builds a Umbraco plugin (package.manifest and css folder) and puts the files in the project folder for Umbraco Plugins.

All paths are configured from package.json.

The build step is disabled by default. But should be enabled in both development and production builds if the project is a Umbraco site.